requirements.txt
asgiref==3.8.1
Django==4.2
django-environ==0.11.2
djangorestframework==3.15.2
packaging==24.1
psycopg2-binary==2.9.9
sqlparse==0.5.1
typing_extensions==4.12.2
whitenoise==6.7.0
